Handle tracks which have no non-nil added date
A nil added-to-playlist date means that the track was added to the
playlist before Spotify kept track of dates added. Previously, the
script could handle a track where _some_ of the dates were nil, but
not where _all_ of them were.

This commit fixes that: if all the dates are nil, then don't count
the track in any date range.