Combine tracks with the same title and artists
Within a chart, combine the playlist adds for duplicate tracks.
These are usually single and album versions of a track. The title
and artists must match exactly, to avoid merging live versions and
remixes with the original track.

Choose a canonical track with the earliest release date so that it
will be consistent when run on different dates.