Choose canonical version of track based on ID, not release date
It would be more consistent to pick the track with the earliest
release date, because the same track would be chosen regardless of
when the script was run. But finding the release date is expensive
because it requires a Spotify API call for every track's album.

Use the track ID instead. This may cause inconsistencies where one
month's chart features the single version of a track, and the next
month's chart features the album version. But the tradeoff seems
worth it if it saves many thousands of API calls.