Fix filtering of playlists which match search term
Add a unit test to ensure that only playlists which exactly match the
search term are used when generating charts. This is because the Spotify
search is not specific enough for our purposes: a search term such as
"wcs" returns playlists with just "wc" in the name and which are
probably not relevant.

Fix the bug caught by this unit test.