Create chart using available versions of songs
Spotify tracks often have several "instances" which may have different
availabilities in different countries. Before, the chart builder picked
a fairly arbitrary instance in a way that was designed to maximise
consistency between charts for different months, but it often chose
tracks which were not playable.

The new logic attempts to find tracks that are available in a wide range
of places:
- Use an instance that is playable in the UK and US if possible. This is
  a reasonable proxy for worldwide availability, because tracks playable
  in the UK are usually available in the rest of Europe. The same goes
  for the US, Cananda and Mexico.
- If that's not possible, find a track that's playable in the UK _or_
  the US. The UK because, selfishly, I live here and I'm mostly building
  this tool for my own amusement. The US because that's where most West
  Coast Swing dancers live, as far as I know.
- If that's not possible, find a track that's available in the maximum
  number of countries.

These are all rough heuristics which I'll revisit if this site ever gets
any kind of international audience.

I'm also considering just using these canonical tracks as a starting
point anyway, and adding multiple versions of tracks to a chart if the
script can't find one which is widely available.