Calculate rising track score based on 3 years of data
Compare the current number of adds to the previous maximum, rather than
just the most recent month. Using just the most recent month is
unsuitable when the data is noisy (which it is) because popular tracks
which are dropping in popularity are still incorrectly included in the
rising track chart when they happen to have a month which is slightly
better than the previous one.

Use the past three years because:
- It's simpler than working out how far back the real data goes
- If an old track starts becoming popular again, the rising track chart
  should reflect that. Popular tracks keep being played for a while, so
  three years seems like a reasonable cut-off (at least as a starting
  point).