Add a cut-off date to rising track chart
The previous algorithm was including too many old tracks. This seems to
be because older Spotify data is incomplete, either because people
delete and re-add tracks to playlists, or because Spotify's playlist
search doesn't find playlists which haven't been updated recently.

Try to compensate for this by ignoring any tracks with adds before three
months ago. This has the danger that it will exclude legitimate rising
tracks which were released a while ago (and were added to just a couple
of playlists by musical pioneers) but became popular more than a few
months later. This might be worth the trade-off of not having Ed
Sheeran's Bloodstream in every month's chart.