Get track's release date from earliest date available
We group tracks by name and artists and choose a canonical track based
on the available markets, but the canonical track may be from a
re-release rather than the original album release so its date may not be
the "true" release date.

Find the earliest avilable date rather the date of the canonical track
to make it easier to group tracks by release date.