Catch 404 error when paging to end of playlist search
It looks like the playlist search hits a 404 response if the number of
playlist search results is a multiple of the batch size, because it
tries to page past the last set of results. In this case, it's OK to
ignore the error response and assume we have reached the last page.