Omit tracks without an ID from the raw track data
There are only a few thousand of these tracks. I think they are tracks
that people have added to playlists from outside of Spotify, and Spotify
can't match them to its own tracks.

At some point, it might be worth matching them on track title and artist
names, but I'll have to rethink how the canonical tracks are stored
because it currently relies on tracks having an ID.

This fixes a bug where all the ID-less tracks were grouped together, and
appeared quite high in the chart under the name of whichever ID-less track
was saved first.